disabled call to check mpi
MPI will be always on if the code is compiled with MPIV preprocessor flag. Note that this is naive solution to prevent mpi hang ups when executing quick.MPI/quick.cuda.MPI with mpirun -np 1. In future, we should either remove all bMPI conditions from the code or put on all mpi code paths.